+++ internal/service/pinpointsmsvoicev2/service_package_gen.go
@@ -82,6 +82,23 @@ func (p *servicePackage) FrameworkResources(ctx context.Context) []*inttypes.Ser
 				WrappedImport: true,
 			},
 		},
+		{
+			Factory:  newSenderIDResource,
+			TypeName: "aws_pinpointsmsvoicev2_sender_id",
+			Name:     "Sender ID",
+			Tags: unique.Make(inttypes.ServicePackageResourceTags{
+				IdentifierAttribute: names.AttrARN,
+			}),
+			Region: inttypes.ResourceRegionDefault(),
+			Identity: inttypes.RegionalParameterizedIdentity([]inttypes.IdentityAttribute{
+				inttypes.StringIdentityAttribute("sender_id", true),
+				inttypes.StringIdentityAttribute("iso_country_code", true),
+			}),
+			Import: inttypes.FrameworkImport{
+				WrappedImport: true,
+				ImportID:      senderIDImportID{},
+			},
+		},
 	}
 }
 
@@ -106,6 +123,19 @@ func (p *servicePackage) FrameworkListResources(ctx context.Context) iter.Seq[*i
 			}),
 			Region:   inttypes.ResourceRegionDefault(),
 			Identity: inttypes.RegionalSingleParameterIdentity(inttypes.StringIdentityAttribute(names.AttrID, true)),
+		},
+		{
+			Factory:  newSenderIDResourceAsListResource,
+			TypeName: "aws_pinpointsmsvoicev2_sender_id",
+			Name:     "Sender ID",
+			Tags: unique.Make(inttypes.ServicePackageResourceTags{
+				IdentifierAttribute: names.AttrARN,
+			}),
+			Region: inttypes.ResourceRegionDefault(),
+			Identity: inttypes.RegionalParameterizedIdentity([]inttypes.IdentityAttribute{
+				inttypes.StringIdentityAttribute("sender_id", true),
+				inttypes.StringIdentityAttribute("iso_country_code", true),
+			}),
 		},
 	})
 }